Rename: #33 ERD 프로젝트 상태 테이블 이름으로 통합 (TaskStatus → ProjectStatus) (#42)
* Rename: #33 ERD 프로젝트 상태 테이블 이름으로 통합 (TaskStatus → ProjectStatus)
