+++ src/types/ProjectStatusType.tsx
@@ -1,13 +1,12 @@
-// ToDo: ERD에 맞춰 ProjectStatus로 변경할 것
 // ToDo: API 설계 완료시 데이터 타입 변경할 것
-export type TaskStatus = {
+export type ProjectStatus = {
   statusId: number;
   name: string;
   color: string;
   order: number;
 };
 
-export type TaskStatusForm = {
+export type ProjectStatusForm = {
   name: string;
   color: string;
 };
