+++ src/types/TaskType.tsx
@@ -1,4 +1,4 @@
-import { TaskStatus } from './TaskStatusType';
+import { ProjectStatus } from '@/types/ProjectStatusType';
 
 // ToDo: API 설계 완료시 데이터 타입 변경할 것
 export type Task = {
@@ -11,4 +11,4 @@ export type Task = {
   endDate: string;
 };
 
-export type TaskWithStatus = TaskStatus & { tasks: Task[] };
+export type TaskWithStatus = ProjectStatus & { tasks: Task[] };
